Fix printing index0 items
We store the index of the item we want to print in the opt_* variable,
then later on we check if we wish to print any item. However, if we wish
to print item 0, the value 0 is stored in opt_* which translates to false
in the if-test. So we don't even trigger the print routine.

To get around this, we add +1 to the opt_* and subtract -1 before we send
it to the print function.